Make world record leaderboard position relative to each user

+++ db.js
@@ -35,7 +35,7 @@ var User = sequelize.define('user', {
 
 var Leaderboard = sequelize.define('leaderboard', {
   username: Sequelize.STRING,
-  time: Sequelize.INTEGER
+  time: Sequelize.FLOAT
 }, {freezeTableName:true});
 
 var Gamestate = sequelize.define('gamestate', {
